+++ modules/javafx/jvm/src/test/scala/intaglio/javafx/JavaFxConformanceSuite.scala
@@ -61,7 +61,7 @@ class JavaFxConformanceSuite extends munit.FunSuite:
               ) && h == horizontal && v == vertical && (rotation != 0.0) == rotated
             case _ => false
           }
-        case RenderRequirement.TextStyle(name, color, fontSizePx, fontFamily, alpha) =>
+        case RenderRequirement.TextStyle(name, color, fontSizePx, fontFamily, alpha, fontWeight) =>
           out.commands.exists {
             case JavaFxCommand.Text(
                   _,
@@ -76,7 +76,8 @@ class JavaFxConformanceSuite extends munit.FunSuite:
                   commandName
                 ) =>
               commandName.contains(name) && paint.fill.contains(JavaFxColor.fromRgba(color)) &&
-              actualFontSize == fontSizePx && actualFontFamily == fontFamily && paint.opacity == alpha
+              actualFontSize == fontSizePx && actualFontFamily == fontFamily &&
+              paint.opacity == alpha && paint.fontWeight == fontWeight
             case _ => false
           }
         case RenderRequirement.Image(name, dimensions, interpolation, alpha) =>